# BackEnd/ADMIN_DA3/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/baiviet/search": {
            "post": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "f4b6a4fc9e7ea6c2f475c2a8d625d87e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "totalPage",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/baiviet/": {
            "get": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "ce5cdfcd6ad4dcb55b7bab5dd8501301",
                "parameters": [
                    {
                        "name": "total",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/baiviet/delete/{id}": {
            "delete": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "c5c57deebda54e4fb5257a10c54e03f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/baiviet/deletes": {
            "delete": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "0a54c0803e4d5faa0a11361da7393667",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/baiviet/save/{id}": {
            "post": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "03acb0071a0efd8531e2806245c20418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "MaTaiKhoan",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "MaChuDe",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "NoiDung",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/baiviet/get/{id}": {
            "get": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Get logged-in user details",
                "operationId": "ffbe4e3a71236593db3eb3600f138acb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của bài viết",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}